winstorecompat: allow building with code that needs codeGeneration capability
By default winstorecompat is built without the need for this capability.

VirtualProtect may be mapped to VirtualProtectFromApp but only if the app has
the codeGeneration capability. It is needed when compiling UNIX code with
relocatable sections that are relocated at runtime.

VirtualProtectFromApp is only available if compiled to target win10.

https://docs.microsoft.com/en-us/windows/win32/api/memoryapi/nf-memoryapi-virtualprotectfromapp
https://docs.microsoft.com/en-us/windows/uwp/packaging/app-capability-declarations