winstorecompat: use the include from the prefix
This should be the folder where the built mingw64 headers are installed when
winstorecompat is built from the same source tree.

This requires to use a "make install" of the headers before building
winstorecompat.

It's still possible to provide another preferred location for the headers via
the CPP_FLAGS of the build environment.