headers: add some windows.storage.accesscache interfaces
IStorageApplicationPermissionsStatics is the interface to use to access the static methods
of StorageApplicationPermissions [1], to use by creating an instance of
"Windows.Storage.AccessCache.StorageApplicationPermissions" with RoGetActivationFactory().

IStorageItemAccessList is described at [2].

StorageApplicationPermissions is not a runtimeclass but a static class with a
few interfaces depending on the contract version. widl doesn't manage this yet
but it's possible to generate similar code (and class string) by declaring it
as a runtimeclass.

[1] https://docs.microsoft.com/en-us/uwp/api/windows.storage.accesscache.storageapplicationpermissions
[2] https://docs.microsoft.com/en-us/uwp/api/windows.storage.accesscache.istorageitemaccesslist