header: add StorageFile/StorageFolder classes to windows.storage
They are not actually runtime classes but sealed classes. The current widl
doesn't understand this keyword so we pretend they are runtime classes. It
generates the same code.

The IStorageFile interface is described at [1].
The IStorageFileStatics is the interface to use to access the static methods
of StorageFile [2], to use by creating an instance of "Windows.Storage.StorageFile"
with RoGetActivationFactory().

[1] https://docs.microsoft.com/en-us/uwp/api/windows.storage.istoragefile
[2] https://docs.microsoft.com/en-us/uwp/api/windows.storage.storagefile