widl: fix delegate class generation
The is_object(stmt->u.type) is asserting because it's created from the delegate
and not the delegate interface. So we allow creating interfaced for delegates
as well.

To create the delegate interface we need the c_name and the uuid attribute.

The created interface has the name of the delegate prefixed with 'I'. This is
in line with the other MS winrt headers.