widl: handle parameterized C++ templates
The C++ templates need to be defined before the actual classes/interfaces can
use them so the definition and declaration is separated.

runtimeclass can be used. The default interface is used to generate the output
is some templates (see AggregateType which can tell between the runtime type
and the default interface in C++ templates).

Multiple parameters, as in ITypedEventHandler are also supported.

The code generated is compatible with code that compiles with MSVC in C++.